requirements.txt
django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
pytest==8.1.1
pytest-django==4.8.0
coverage==7.4.4
psycopg2-binary==2.9.9